fix sync_read_message reading too much data
While implementing the Gossip module and while testing with the `gossip_client.py` I came across the following issue:

```
[+] Connected to gossip module: ('127.0.0.1', 7001)
[+] Sent GOSSIP_NOTIFY for type 1337).
[+] Waiting for GOSSIP_NOTIFICATION message...
[+] Got GOSSIP_NOTIFICATION: mID = 34250, type = 1337, data = b'deadbeef\x00\x10'
00000000: 00 10 01 F6 85 CA 05 39  64 65 61 64 62 65 65 66  .......9deadbeef
00000010: 00 10                                             ..
[+] Sent GOSSIP_VALIDATION: mid = 34250, Valid).
[+] Waiting for GOSSIP_NOTIFICATION message...
[-] Wrong packet type: 51124:
00000000: 01 F6 C7 B4 05 39 64 65  61 64 62 65 65 66        .....9deadbeef
[-] Exiting.
```

Here the client read too much data (the header specified 16 bytes) and therefore fails to process the next message correctly. This is caused by sync_read_message reading two bytes for the message length and then reading the specified number of bytes. This is invalid as the size specified in the header includes the header itself and therefore we should only further read the specified size minus two bytes in total.